refactor(ui): colocate loading skeletons and close remaining gaps
Split the PageSkeletons megafile into route/domain modules so anatomy
lives next to the pages that use it. Add invite + referral-code route
loaders, and replace leftover client spinner/text loading with matching
skeletons.
