+++ app/checkout/checkout-success/page.tsx
@@ -5,7 +5,7 @@ import { useEffect, useState, Suspense } from "react";
 import { useSearchParams, useRouter } from "next/navigation";
 import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
 import { Button } from "@/components/ui/button";
-import { CheckoutResultSkeleton } from "@/components/loading/PageSkeletons";
+import { CheckoutResultSkeleton } from "@/app/checkout/CheckoutSkeletons";
 import { CheckCircle, Clock, Calendar, ArrowRight } from "lucide-react";
 interface PaymentDetails {
   paymentIntent: string;
@@ -163,11 +163,7 @@ function CheckoutSuccessContent() {
   };
 
   if (loading) {
-    return (
-      <div className="min-h-screen flex items-center justify-center bg-muted">
-        <div className="animate-spin rounded-full h-8 w-8 border-b-2 border-foreground"></div>
-      </div>
-    );
+    return <CheckoutResultSkeleton />;
   }
 
   if (!paymentDetails) {
