+++ app/dashboard/admin/analytics/loading.tsx
@@ -1,4 +1,4 @@
-import { AnalyticsSkeleton } from "@/components/loading/PageSkeletons";
+import { AnalyticsSkeleton } from "@/components/dashboard/DashboardSkeletons";
 
 export default function Loading() {
   return <AnalyticsSkeleton />;
