+++ app/dashboard/consultant/[consultantId]/(features)/requests/[requestId]/allocate/loading.tsx
@@ -1,4 +1,4 @@
-import { HeatmapSkeleton } from "@/components/loading/PageSkeletons";
+import { HeatmapSkeleton } from "@/components/scheduling/CalendarSkeletons";
 
 export default function Loading() {
   return <HeatmapSkeleton />;
