+++ app/dashboard/consultee/[consulteeId]/(features)/appointments/[appointmentId]/reschedule/loading.tsx
@@ -1,4 +1,4 @@
-import { HeatmapSkeleton } from "@/components/loading/PageSkeletons";
+import { HeatmapSkeleton } from "@/components/scheduling/CalendarSkeletons";
 
 export default function Loading() {
   return <HeatmapSkeleton />;
