+++ app/dashboard/organization/[orgId]/appointments/[appointmentId]/loading.tsx
@@ -1,4 +1,4 @@
-import { AppointmentDetailSkeleton } from "@/components/loading/PageSkeletons";
+import { AppointmentDetailSkeleton } from "@/components/dashboard/DashboardSkeletons";
 
 export default function Loading() {
   return <AppointmentDetailSkeleton />;
