+++ app/explore/activities/loading.tsx
@@ -1,4 +1,4 @@
-import { MarketingPageSkeleton } from "@/components/loading/PageSkeletons";
+import { MarketingPageSkeleton } from "@/components/loading/MarketingPageSkeleton";
 
 export default function Loading() {
   return <MarketingPageSkeleton />;
