+++ app/explore/enterprise/organisations/loading.tsx
@@ -1,4 +1,4 @@
-import { OrganisationsExploreSkeleton } from "@/components/loading/PageSkeletons";
+import { OrganisationsExploreSkeleton } from "./OrganisationsExploreSkeleton";
 
 export default function Loading() {
   return <OrganisationsExploreSkeleton />;
