+++ app/explore/programs/plans/subscriptions/[subscriptionPlanId]/loading.tsx
@@ -1,4 +1,4 @@
-import { PlanDetailSkeleton } from "@/components/loading/PageSkeletons";
+import { PlanDetailSkeleton } from "../../PlanDetailSkeleton";
 
 export default function Loading() {
   return <PlanDetailSkeleton />;
