+++ app/meetings/[id]/loading.tsx
@@ -1,4 +1,4 @@
-import { MeetingRoomSkeleton } from "@/components/loading/PageSkeletons";
+import { MeetingRoomSkeleton } from "./MeetingRoomSkeleton";
 
 export default function Loading() {
   return <MeetingRoomSkeleton />;
