+++ app/organizations/invite/[token]/loading.tsx
@@ -1,0 +1,5 @@
+import { InviteAcceptSkeleton } from "./InviteAcceptSkeleton";
+
+export default function Loading() {
+  return <InviteAcceptSkeleton />;
+}
