+++ app/r/[code]/loading.tsx
@@ -1,7 +1,14 @@
+import { Skeleton } from "@/components/ui/skeleton";
+
 export default function Loading() {
   return (
-    <div className="flex min-h-screen items-center justify-center">
-      <div className="h-10 w-10 animate-spin rounded-full border-4 border-muted border-t-foreground" />
-    </div>
+    <main className="flex min-h-screen items-center justify-center bg-background p-6">
+      <div className="w-full max-w-sm space-y-4 rounded-xl border border-border bg-card p-8 text-center">
+        <Skeleton className="mx-auto h-12 w-12 rounded-full" />
+        <Skeleton className="mx-auto h-6 w-40" />
+        <Skeleton className="mx-auto h-4 w-56" />
+        <Skeleton className="mx-auto h-10 w-full rounded-lg" />
+      </div>
+    </main>
   );
 }
