+++ components/dashboard/DashboardSkeletons.tsx
@@ -400,3 +400,61 @@ export function PlannerSkeleton() {
     </div>
   );
 }
+
+
+/** Appointment detail: banner + content columns. */
+export function AppointmentDetailSkeleton() {
+  return (
+    <div className="space-y-6 p-6 lg:p-8">
+      <div className="space-y-3 rounded-xl border border-border bg-card p-6">
+        <Skeleton className="h-6 w-32 rounded-full" />
+        <Skeleton className="h-8 w-64" />
+        <Skeleton className="h-4 w-48" />
+        <div className="flex flex-wrap gap-2 pt-2">
+          <Skeleton className="h-10 w-28" />
+          <Skeleton className="h-10 w-28" />
+          <Skeleton className="h-10 w-28" />
+        </div>
+      </div>
+      <div className="grid gap-6 lg:grid-cols-3">
+        <div className="space-y-4 lg:col-span-2">
+          {[1, 2, 3].map((i) => (
+            <Skeleton key={i} className="h-28 rounded-xl" />
+          ))}
+        </div>
+        <div className="space-y-4">
+          <Skeleton className="h-40 rounded-xl" />
+          <Skeleton className="h-32 rounded-xl" />
+        </div>
+      </div>
+    </div>
+  );
+}
+
+/** Admin analytics: 4 stat cards + chart panels. */
+export function AnalyticsSkeleton() {
+  return (
+    <div className="space-y-6 p-6 lg:p-8">
+      <div className="space-y-2">
+        <Skeleton className="h-8 w-40" />
+        <Skeleton className="h-4 w-64" />
+      </div>
+      <div className="grid grid-cols-1 gap-4 md:grid-cols-2 lg:grid-cols-4">
+        {[1, 2, 3, 4].map((i) => (
+          <div
+            key={i}
+            className="space-y-3 rounded-xl border border-border bg-card p-5"
+          >
+            <Skeleton className="h-4 w-24" />
+            <Skeleton className="h-8 w-20" />
+            <Skeleton className="h-3 w-16" />
+          </div>
+        ))}
+      </div>
+      <div className="grid gap-6 lg:grid-cols-2">
+        <Skeleton className="h-72 rounded-xl" />
+        <Skeleton className="h-72 rounded-xl" />
+      </div>
+    </div>
+  );
+}
